build: refactor :network:base and :network:impl modules
