+++ app/src/main/java/com/alanvo/androiddemo/di/AppModules.kt
@@ -1,9 +1,12 @@
 package com.alanvo.androiddemo.di
 
+import com.vnhanh.network.impl.di.networkModule
+
 
 val appModules = listOf(
+    networkModule,
     viewModelsModule,
     useCasesModule,
     repositoriesModule,
-    apisModule
+    apisModule,
 )
