+++ network/base/src/main/java/com/vnhanh/demo/network/base/common/INetworkChecker.kt
@@ -1,4 +1,4 @@
-package com.vnhanh.demo.network.base
+package com.vnhanh.demo.network.base.common
 
 interface INetworkChecker {
     fun isInternetAvailable(): Boolean
