+++ network/base/src/main/java/com/vnhanh/demo/network/base/data/IApiResponse.kt
@@ -1,4 +1,4 @@
-package com.vnhanh.demo.network.base
+package com.vnhanh.demo.network.base.data
 
 
 /**
