+++ network/impl/src/main/java/com/vnhanh/network/impl/di/NetworkModule.kt
@@ -1,9 +1,18 @@
 package com.vnhanh.network.impl.di
 
+import com.vnhanh.demo.network.base.serviceCreator.IApiServiceCreator
+import com.vnhanh.network.impl.httpClient.ApiServiceCreatorImpl
+import com.vnhanh.network.impl.httpClient.AppOkhttpClient
+import okhttp3.Interceptor
 import org.koin.dsl.module
 
 val networkModule = module {
-    single {
 
+    single { (curlTag: String, interceptors: List<Interceptor>, isDebugMode: Boolean) ->
+        AppOkhttpClient.getBuilder(curlTag, interceptors, isDebugMode)
+    }
+
+    single<IApiServiceCreator> {
+        ApiServiceCreatorImpl(get())
     }
 }
